Properly avoid overflows of width and height and cast integers to 64-bit where necessary to do proper calculations.
git-svn-id: svn://svn.netsurf-browser.org/trunk@6254 d004b94f-0d10-0410-95b5-d60e4a4a7a9a

+++ libnsbmp/libnsbmp.h
@@ -70,8 +70,8 @@ typedef struct bmp_bitmap_callback_vt_s {
 typedef struct bmp_image {
 	bmp_bitmap_callback_vt bitmap_callbacks;	/**< callbacks for bitmap functions */
 	uint8_t *bmp_data;				/** pointer to BMP data */
-	uint16_t width;					/** width of BMP (valid after _analyse) */
-	uint16_t height;				/** heigth of BMP (valid after _analyse) */
+	uint32_t width;					/** width of BMP (valid after _analyse) */
+	uint32_t height;				/** heigth of BMP (valid after _analyse) */
 	bool decoded;					/** whether the image has been decoded */
 	void *bitmap;					/** decoded image */
 	/**	Internal members are listed below
